refactoring/cleaning DataCite provider, drop cache
